requirements.txt
Django==1.6.5
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-registration==1.0
django-tastypie==0.11.0
django-tastypie-swagger==0.1.2
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.6.1
static==1.0.2
wsgiref==0.1.2
requests==2.3.0
beautifulsoup4==4.3.2
